% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hagel.R
\docType{package}
\name{hagel}
\alias{hagel}
\title{Download weather radar data for Belgium and The Netherlands}
\description{
Download weather radar data for Belgium and The Netherlands.
}
\author{
Pieter Provoost, \email{pieterprovoost@gmail.com}
}
