% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi.R
\docType{data}
\name{knmi_datasets}
\alias{knmi_datasets}
\title{List of KNMI datasets.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.
}
\usage{
knmi_datasets
}
\value{
Tibble with dataset names and versions.
}
\description{
List of KNMI datasets.
}
\keyword{datasets}
