% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi.R
\name{knmi_to_polygon}
\alias{knmi_to_polygon}
\title{Convert KNMI radar image to sfc polygons.}
\usage{
knmi_to_polygon(knmi_raster, threshold = 50, method = "chaikin", ...)
}
\arguments{
\item{knmi_raster}{KNMI image.}

\item{method}{smoothing method.}
}
\value{
sfc.
}
\description{
Convert KNMI radar image to sfc polygons.
}
