% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi.R
\name{knmi_urls}
\alias{knmi_urls}
\title{Get KNMI file URLs.}
\usage{
knmi_urls(
  dataset_name = "radar_reflectivity_composites",
  dataset_version = "2.0",
  after = NULL,
  max_keys = 12
)
}
\arguments{
\item{dataset_name}{dataset name.}

\item{dataset_version}{dataset version.}

\item{after}{return images dated after this time (POSIXct), if not provided then defaults to one hour ago.}

\item{max_keys}{maximum number of keys to return.}
}
\value{
Image URLs.
}
\description{
Get KNMI file URLs.
}
