% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi.R
\name{radar_knmi}
\alias{radar_knmi}
\title{Read radar images from KNMI.}
\usage{
radar_knmi(
  dataset_name = "radar_reflectivity_composites",
  dataset_version = "2.0",
  after = NULL,
  max_keys = 12
)
}
\arguments{
\item{dataset_name}{dataset name.}

\item{dataset_version}{dataset version.}

\item{after}{return images dated after this time (POSIXct), if not provided then defaults to one hour ago.}

\item{max_keys}{maximum number of keys to return.}
}
\value{
List of stars objects.
}
\description{
Read radar images from KNMI.
}
