% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi.R
\name{read_knmi_file}
\alias{read_knmi_file}
\title{Read KNMI radar file.}
\usage{
read_knmi_file(
  filename,
 
    radar_proj = "+proj=stere +lat_0=90 +lon_0=0 +lat_ts=60 +a=6378140 +b=6356750 +x_0=0 y_0=0"
)
}
\arguments{
\item{filename}{local filename.}

\item{radar_proj}{proj string of the radar file.}
}
\value{
A stars object.
}
\description{
Read KNMI radar file.
}
